test: handle the case when ObjectStore::create returns NULL
When the objectstore type doesn't exist, NULL is returned in SetUp.
Handle the NULL return code to avoid a segment fault.

+++ src/test/objectstore/store_test.cc
@@ -55,6 +55,10 @@ public:
                                               string(GetParam()),
                                               string("store_test_temp_dir"),
                                               string("store_test_temp_journal"));
+    if (!store_) {
+      cerr << __func__ << ": objectstore type " << string(GetParam()) << " doesn't exist yet!" << std::endl;
+      return;
+    }
     store.reset(store_);
     EXPECT_EQ(store->mkfs(), 0);
     EXPECT_EQ(store->mount(), 0);
